WIP. Implementing V2 RPCs.
